2025CR468	6/16/2025	MikeR	Feature	Add MetaScore Table and MetaScoreType 2025CR469	6/16/2025	MikeR	Feature	Track metascores for all reporting (check for them) 2025CR470	6/16/2025	MikeR	Feature	Modify AssessmentResults XML to include metascores. 2025CR473	6/17/2025	MikeR	Feature	Add includeAiScore flag to reports 2025CR474	6/17/2025	MikeR	Feature	Add EvauationPlanId (jobId) to testkey and rccheck and lvi 2025CR475	6/17/2025	MikeR	Feature	Add Report Flag "skipaiscoressection" to hide AI Scores sections.

+++ src/main/java/com/tm2score/av/AvScoringUtils.java
@@ -431,7 +431,7 @@ public class AvScoringUtils {
         Integer o = (Integer) cout[2];
         
         if( o!=null && o.intValue()!=0 )
-            uuf.setOrientation( o.intValue() );
+            uuf.setOrientation(o);
 
         // LogService.logIt( "AvScoringUtils.checkVideoOrientation() CCC Orietnation=" + (o!=null ? o.toString() : "null") );
